+++ src/crispy_forms_gds/layout/fields.py
@@ -34,7 +34,7 @@ class Field(crispy_forms_layout.LayoutObject):
 
     @classmethod
     def checkboxes(
-        cls, field, legend_size=None, legend_tag=None, small=False, **kwargs
+        cls, field, legend_size=None, legend_tag=None, small=False, hints=None, **kwargs
     ):
         """
         Create a field for displaying checkboxes.
@@ -49,6 +49,9 @@ class Field(crispy_forms_layout.LayoutObject):
                 Default is None.
 
             small (bool): Display small checkboxes. Default is False.
+
+            hints (dict): Hints that will be displayed for one or more checkboxes.
+                The dictionary key is the checkbox value. Defaults to None.
 
             **kwargs: Attributes to add to the <input> element when the field is
                 rendered.
@@ -66,6 +69,9 @@ class Field(crispy_forms_layout.LayoutObject):
             context["legend_tag"] = legend_tag
 
         context["checkboxes_small"] = small
+
+        if hints:
+            context["checkboxes_hints"] = hints
 
         return Field(field, context=context, **kwargs)